refactor(inductor): rename require layout pass
